+++ src/rectangle.hpp
@@ -3,13 +3,14 @@
 
 #include <cstring> // for size_t
 #include <stdexcept>
+#include "utils.hpp"
 
 using namespace std;
 
 class WorldDimension {
 public:
 
-    WorldDimension(size_t width, size_t height) :
+    WorldDimension(uint32_t width, uint32_t height) :
         _width(width), _height(height)
     {
     };
@@ -25,53 +26,53 @@ public:
         return wd;
     };
 
-    size_t getWidth() const
+    uint32_t getWidth() const
     {
         return _width;
     };
 
-    size_t getHeight() const
+    uint32_t getHeight() const
     {
         return _height;
     };
 
-    size_t getMax() const
+    uint32_t getMax() const
     {
         return _width > _height ? _width : _height;
     }
 
-    size_t getArea() const
+    uint32_t getArea() const
     {
         return _width * _height;
     }
 
-    size_t xMod(size_t x) const
+    uint32_t xMod(uint32_t x) const
     {
         return x % getWidth();
     }
 
-    size_t yMod(size_t y) const
+    uint32_t yMod(uint32_t y) const
     {
         return y % getHeight();
     }    
 
-    bool contains(const size_t x, const size_t y) const
+    bool contains(const uint32_t x, const uint32_t y) const
     {
         return (x >= 0 && x < _width && y >= 0 && y < _height);
     }
 
-    void normalize(size_t& x, size_t& y) const
+    void normalize(uint32_t& x, uint32_t& y) const
     {
         x %= _width;
         y %= _height;
     }
 
-    size_t indexOf(const size_t x, const size_t y) const
+    uint32_t indexOf(const uint32_t x, const uint32_t y) const
     {
         return y * getWidth() + x;
     }
 
-    size_t lineIndex(const size_t y) const
+    uint32_t lineIndex(const uint32_t y) const
     {
         if (y<0 || y>=_height){
             throw invalid_argument("WorldDimension::line_index: y is not valid");
@@ -79,40 +80,40 @@ public:
         return indexOf(0, y);
     }
 
-    size_t yFromIndex(const size_t index) const
+    uint32_t yFromIndex(const uint32_t index) const
     {
         return index / _width;
     }
 
-    size_t xFromIndex(const size_t index) const
+    uint32_t xFromIndex(const uint32_t index) const
     {
-        const size_t y = yFromIndex(index);
+        const uint32_t y = yFromIndex(index);
         return index - y * _width;
     }    
 
-    size_t normalizedIndexOf(const size_t x, const size_t y) const
+    uint32_t normalizedIndexOf(const uint32_t x, const uint32_t y) const
     {
         return indexOf(xMod(x), yMod(y));
     }
 
-    size_t xCap(const size_t x) const
+    uint32_t xCap(const uint32_t x) const
     {
         return x < _width ? x : (_width-1);
     }
 
-    size_t yCap(const size_t y) const
+    uint32_t yCap(const uint32_t y) const
     {
         return y < _height ? y : (_height-1);
     }
 
-    size_t largerSize() const
+    uint32_t largerSize() const
     {
         return _width > _height ? _width : _height;
     }
 
 private:
-    const size_t _width;
-    const size_t _height;
+    const uint32_t _width;
+    const uint32_t _height;
 };
 
 namespace Platec {
@@ -120,8 +121,8 @@ namespace Platec {
 class Rectangle {
 public:
     Rectangle(const WorldDimension& worldDimension,
-            size_t left, size_t right,
-            size_t top, size_t bottom)
+            uint32_t left, uint32_t right,
+            uint32_t top, uint32_t bottom)
             : _worldDimension(worldDimension),
               _left(left), _right(right),
               _top(top), _bottom(bottom)
@@ -142,50 +143,50 @@ public:
     };
 
 
-    size_t getMapIndex(size_t* px, size_t* py) const throw();
-    void enlarge_to_contain(size_t x, size_t y);
+    uint32_t getMapIndex(uint32_t* px, uint32_t* py) const throw();
+    void enlarge_to_contain(uint32_t x, uint32_t y);
 
-    size_t getLeft() const
+    uint32_t getLeft() const
     {
         return _left;
     }
 
-    size_t getRight() const
+    uint32_t getRight() const
     {
         return _right;
     }
 
-    size_t getTop() const
+    uint32_t getTop() const
     {
         return _top;
     }
 
-    size_t getBottom() const
+    uint32_t getBottom() const
     {
         return _bottom;
     }
 
-    void setLeft(size_t v)
+    void setLeft(uint32_t v)
     {
         _left = v;
     }
 
-    void setRight(size_t v)
+    void setRight(uint32_t v)
     {
         _right = v;
     }
 
-    void setTop(size_t v)
+    void setTop(uint32_t v)
     {
         _top = v;
     }
 
-    void setBottom(size_t v)
+    void setBottom(uint32_t v)
     {
         _bottom = v;
     }
 
-    void shift(size_t dx, size_t dy)
+    void shift(uint32_t dx, uint32_t dy)
     {
         _left   += dx;
         _right  += dx;
@@ -195,8 +196,8 @@ public:
 
 private:
     const WorldDimension _worldDimension;
-    size_t _left, _right;
-    size_t _top, _bottom;
+    uint32_t _left, _right;
+    uint32_t _top, _bottom;
 };
 
 };
